+++ packages/asc-ui/src/components/Typography/TypographyStyle.ts
@@ -5,14 +5,12 @@ import TypographyElements = Theme.TypographyElements
 
 export type Props = {
   gutterBottom?: number
-  paragraph?: boolean
   element?: Variant
   color?: Theme.ColorType
   fontSize?: number
   styleAs?: keyof TypographyElements
   as?: any
-  strong?: boolean // makes the text strong(bold)
-  compact?: boolean // applies the compact line heights
+  strong?: boolean
 }
 
 export const defaultTypographyStyles = {
@@ -46,5 +44,5 @@ export default styled.p<Props>`
     fontSize &&
     css`
       font-size: ${fontSize}px;
-    `} 
+    `}
 `