+++ packages/asc-ui/src/components/CustomHTMLBlock/CustomHTMLBlockStyle.ts
@@ -37,6 +37,5 @@ export default styled.div`
   video {
     max-width: 100%;
     margin: 24px auto;
-    width: 100%;
   }
 `