+++ CHANGELOG.md
@@ -16,6 +16,9 @@ Prefix the change with one of these keywords:
 
 ## UNRELEASED
 
+- Removed: specific parameter hasLongText from the paragraph, will be implemented in the consumer applications
+
+
 ## [0.15.0] - 2019-09-17
 
 - Added: theme spacing unit and a themeSpacing util to retrieve a size from the theme's spacing unit.